fix: rename filter_routes.py to filter_utils.py and move it to utils folder
